Saving current pokemon status
In this commit I started to add the serialization. Using a save button,
is possible to generate a serialized file named "user.dat" with the data
of the current pokemon status and reload it in the next time the program
is started. I implemented some other classes to make the serialized data
more close to the final data that I want to store in the final fase of
the project, for that, I created the classes "personagem"(character),
"item"  besides pokeball and potion which extend the item class