Keyboard Arrows, FighFrame, Character Name
The character now can be moved by the keyboard arrows, the items now
aren't infinite anymore during the battles. The pokemons defeated change
places automatically with one alive, the pokemons defeated can't return
to the battle but at the end of the battle all of the pokemons in the
team should be reseted. In the choose panel you have to give your
character a name.